refactor(gui): エディタ共通のファイル操作を BaseEditorWindow に集約
4つのエディタで重複していた _on_open / _on_save / _on_save_as を
BaseEditorWindow 基底クラスに抽出し, 各エディタは _file_filter と
open_path / save_to を定義するだけで済むようにした.
